move the vite plugin to dashboard-core
