+++ packages/dashboard-core/package.json
@@ -11,6 +11,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./plugin": "./src/plugin.ts",
+    "./vite": "./src/vite/index.ts",
     "./providers/*": "./src/providers/*.tsx",
     "./hooks/*": "./src/hooks/*.ts",
     "./hooks/use-command-palette": "./src/hooks/use-command-palette.tsx",
@@ -45,6 +46,7 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
+    "@tailwindcss/vite": "^4.2.4",
     "lucide-react": "^1.14.0",
     "spark-md5": "^3.0.2",
     "zod": "^4.4.3"
