+++ packages/dashboard/package.json
@@ -68,7 +68,6 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.59.1",
-    "@tailwindcss/vite": "^4.2.4",
     "@tanstack/router-plugin": "^1.167.35",
     "@types/node": "^25.6.0",
     "@types/react": "^19.2.14",
